refactor: rename nodes files to .service format
